remove clear ioapic
Remove call to clear_IO_APIC in early boot

In hyper-v we can expect the IOAPIC to be empty on boot. The call
to clear_IO_APIC is then not necessary or desirable because it can
cause significant boot delays in hyper-v. We were seeing up to 300 ms
delays with this call on 32-core systems.

Link: https://github.com/clearlinux-pkgs/linux-hyperv/raw/5.6.17-298/0124-remove-clear-ioapic.patch
Signed-off-by: Nathan Chancellor <nathan@kernel.org>
Signed-off-by: Kazuki Hashimoto <kaz205@tuta.io>