HACK: apple-soc-cpufreq: Return some random freq when it isn't found
When I remove certain pstates in the dts, I get mountains of errors saying
that it could not find the frequency for a certain pstate. This occurs
because Apple uses the highest pstate during boot despite the kernel
being unaware of these. Return some random frequency in that scenario to
not break cpufreq.

Signed-off-by: Kazuki H <kazukih0205@gmail.com>